Move instantiation code into its own file
Also do a bit of cleanup around import resolution.
